feat(ui): create auth system

+++ apps/ui/components/auth/withAuth.tsx
@@ -1,11 +1,16 @@
 import { notification, Spin } from 'antd';
 import { useRouter } from 'next/router';
 import { useEffect, useState } from 'react';
-import { NextPageWithLayout } from '@/utils/index';
 import HomeLayout from '@/layout/HomeLayout';
+import { NextPageWithLayout } from '@/utils/types';
+import { useAuthState } from '@/hooks/auth/mutation.hook';
+import config from '@/config/App';
+import { AppRoutes } from '@/utils/AppRoutes';
+import { useAuthUser } from '@/hooks/auth/query.hook';
 
-const deleteAllCookies = () => {
+export const deleteAllCookies = () => {
   const cookies = document.cookie.split(';');
+  console.log(cookies);
   cookies.forEach(function (c) {
     document.cookie = c
       .replace(/^ +/, '')
@@ -15,67 +20,47 @@ const deleteAllCookies = () => {
 
 export function withAuth<P>(
   WrappedComponent: NextPageWithLayout,
-  permissions?: string[],
   isPublic = false
 ) {
   const ComponentWithPermissions = (props: P) => {
     const router = useRouter();
+    const { refetch, isFetching } = useAuthUser();
     const [loading, setLoading] = useState<boolean>(false);
-    /*     const [GetAuthUserQuery] = useGetAuthUserLazyQuery();
-    const user = useSnapshot(AuthState).user as User;
+    const [user, setUser] = useAuthState(null);
+    const getLayout = WrappedComponent.getLayout;
 
     useEffect(() => {
-      const logout = () => {
-        localStorage.clear();
+      if (isPublic || user) {
+        setLoading(false);
+        return;
+      }
+      setLoading(true);
+
+      if (!config.JWT_TOKEN) {
         deleteAllCookies();
-        AuthActions.setUser(null);
-        router.push(AppRoutes.SignIn);
-      };
-      const jwt = localStorage.getItem(config.jwtName);
-      if (!jwt && !isPublic) {
         notification.error({
           message: 'لا يمكنك الوصول لهذه الصفحة!',
           description: 'يرجى تسجيل دخولك لتتمكن من الوصول لخدمات الموقع.',
         });
+
         router.push(AppRoutes.SignIn);
         return;
       }
-
-      if (jwt && !user) {
-        GetAuthUserQuery()
-          .then(({ data }) => {
-            if (!data?.getAuthUser) {
-              logout();
-              return;
-            }
-            const u = data.getAuthUser as User;
-            AuthActions.setUser(u);
-          })
-          .catch((error) => {
-            logout();
-            Logger.log(error);
-          })
-          .finally(() => {
-            setLoading(false);
-          });
-      } else {
-        if (
-          permissions &&
-          !user.role.permissions.some((el) => permissions.includes(el.title))
-        ) {
-          router.push('/');
-          notification.warning({
-            message: 'لا يمكنك الوصول لهذه الصفحة!',
-            description: 'لا تمتلك الصلاحيات الكافية للوصول لهذه الصفحة.',
-          });
-          return;
-        }
-        setLoading(false);
+      if (!user) {
+        refetch().then(({ isError, data }) => {
+          if (isError) {
+            localStorage.clear();
+            deleteAllCookies();
+            router.push(AppRoutes.SignIn);
+          }
+          if (data) {
+            setUser(data);
+          }
+        });
       }
-    }, [GetAuthUserQuery, router, user]); */
-
-    const getLayout = WrappedComponent.getLayout;
-    return loading ? (
+      setLoading(false);
+    }, []);
+    return loading || isFetching ? (
       <HomeLayout>
         <Spin style={{ display: 'block', margin: '5em auto' }} size="large" />
       </HomeLayout>
